server/seed-scripts/requirements.txt
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
pandas==2.3.3
psycopg2-binary==2.9.11
